organize codebase
